pyproject.toml
[project]
name = "eventer"
version = "0.1.0"
description = ""
authors = [
    {name = "Zafk1eI",email = "sintfrank@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11, <4.0.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "sqlalchemy[asyncio] (>=2.0.39,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "fastapi-crudrouter (>=0.8.6,<0.9.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "fastcrud (>=0.15.7,<0.16.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "fastapi-keycloak-middleware (>=1.2.0,<2.0.0)",
    "python-keycloak (>=5.3.1,<6.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "minio (>=7.2.15,<8.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "pillow (>=11.2.1,<12.0.0)"
]

[tool.poetry.scripts]
dev = "uvicorn src.eventer.main:app --host 127.0.0.1 --port 8000 --reload"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
